+++ dsapy/gui/sorting_visualizer/sorting_model.py
@@ -1,9 +1,11 @@
 import random
-from .sorting_config import SIZE_CONFIG, ALGORITHM_CONFIG
+from .sorting_config import SIZE_CONFIG
 from ...src.algorithms.sorting.bubble_sort import BubbleSort
 from ...src.algorithms.sorting.insertion_sort import InsertionSort
+from ...src.algorithms.sorting.merge_sort import MergeSort
 from ...src.algorithms.sorting.quick_sort import QuickSort
 from ...src.algorithms.sorting.selection_sort import SelectionSort
+from ...src.algorithms.sorting.sorting_algorithm import SortingAlgorithm
 
 
 class SortingModel:
@@ -77,7 +79,7 @@ class SortingModel:
         return self.__metadata
 
     @staticmethod
-    def __select_algorithm(algorithm: str) -> BubbleSort | InsertionSort | SelectionSort | QuickSort:
+    def __select_algorithm(algorithm: str) -> SortingAlgorithm:
         """
         Returns the sorter object according to the algorithm input
 
@@ -93,6 +95,8 @@ class SortingModel:
                 return SelectionSort()
             case 'Quick Sort':
                 return QuickSort()
+            case 'Merge Sort':
+                return MergeSort()
             case _:
                 raise ValueError('Unexpected Algorithm')